move fillna, update, where
